refactor: split AppServiceProvider into focused providers
Move Passport configuration (scopes, tokens, expiration, auth view) into
PassportServiceProvider where it belongs alongside the server setup.

Move rate limiter definitions into RouteServiceProvider::configureRateLimiting()
following Laravel's conventional placement for request throttling.

Remove Schema::defaultStringLength(191) which is unnecessary with utf8mb4
on MySQL 5.7.7+ (the minimum for Laravel 12).

Remove Paginator::useBootstrap() as the frontend will be replaced.

AppServiceProvider now contains only: observers, Horizon/Pulse auth gates,
URL configuration, and validator settings.
